Bot/requirements.txt
discord.py
aiofiles
python-dateutil
aiohttp
